django-backend/requirements.txt
appnope==0.1.0
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
djangorestframework==3.10.3
entrypoints==0.3
httpie==1.0.3
idna==2.8
importlib-metadata==0.23
influxdb==5.2.3
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.3.1
urllib3==1.25.6
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
zipp==0.6.0
